Django REST Framework/Work_Flow/snippets/urls.py
from django.urls import path, include
from rest_framework.urlpatterns import format_suffix_patterns
from snippets import views


urlpatterns = [
    # This is before auth
    path('snippets/', views.SnippetList.as_view()),
    path('snippets/<int:pk>/', views.SnippetDetail.as_view()),
    # This is after auth
    path('users/', views.UserList.as_view()),
    path('users/<int:pk>/', views.UserDetail.as_view()),
]

urlpatterns += [
    path('api-auth/', include('rest_framework.urls')),
]

urlpatterns = format_suffix_patterns(urlpatterns)
