Django REST Framework/Work_Flow/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
